refactor(rubrics): Use a single file for all domain errors
