--- src/rubrics/domain/errors/rubric_not_found_error.go
@@ -1,13 +1,0 @@
-package errors
-
-import "net/http"
-
-type RubricNotFoundError struct{}
-
-func (err *RubricNotFoundError) Error() string {
-	return "Rubric not found"
-}
-
-func (err *RubricNotFoundError) StatusCode() int {
-	return http.StatusNotFound
-}
